requirement.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
click==7.1.2
dataclasses==0.8
dj-database-url==0.5.0
Django==3.0
django-bootstrap3==15.0.0
importlib-metadata==2.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
zipp==3.4.1
